api/requirements.txt
tensorflow-gpu==1.14
scipy==1.3.3
numpy==1.14.5
Pillow==6.2.1
requests==2.22.0
keras==2.3.1
tqdm
Flask==0.12.2
flask-cors
gunicorn==19.9.0
dlib
opencv-python
torchvision